feat: Updated 'with related...' to be refactored article instead
