+++ packages/provider/__tests__/article-provider.test.js
@@ -19,7 +19,7 @@ const renderComponent = child =>
   renderer.create(
     <MockedProvider mocks={mocks}>
       <ArticleProvider
-        id="3107c018-cb60-11e4-81dd-064fe933cd41"
+        id="113e9875-b7bf-4dd7-ac99-dee231bf6e74"
         debounceTimeMs={0}
       >
         {child}
