+++ packages/provider/article.js
@@ -1,7 +1,7 @@
 import gql from "graphql-tag";
 import connectGraphql from "./connect";
 
-export const query = gql`
+export const query = gql `
   query ArticleQuery($id: ID!) {
     article(id: $id) {
       id
@@ -23,6 +23,44 @@ export const query = gql`
             ratio
             url
           }
+        }
+      }
+      relatedArticles {
+        id
+        headline
+        byline
+        label
+        publicationName
+        publishedTime
+        summary100: summary( maxCharCount: 100)
+        summary110: summary( maxCharCount: 110)
+        summary125: summary( maxCharCount: 125)
+        summary150: summary( maxCharCount: 150)
+        summary180: summary( maxCharCount: 180)
+        leadAsset {
+          ... on Image {
+            title
+            crop169: crop(ratio: "16:9") {
+              url
+            }
+            crop32: crop(ratio: "3:2") {
+              url
+            }
+          }
+        }
+        url
+      }
+      relatedArticlesLayout {
+        ... on Default {
+          template
+        }
+        ... on LeadAndTwo {
+          template
+          lead
+        }
+        ... on OpinionAndTwo {
+          template
+          opinion
         }
       }
     }
