+++ packages/provider/provider.js
@@ -3,7 +3,6 @@ import AuthorProfileProvider from "./author-profile";
 import AuthorArticlesNoImagesProvider from "./author-articles-no-images";
 import AuthorArticlesWithImagesProvider from "./author-articles-with-images";
 import ArticleProvider from "./article";
-import ArticleRelatedProvider from "./article-with-related-articles";
 
 export default connect;
 
@@ -11,6 +10,5 @@ export {
   AuthorProfileProvider,
   AuthorArticlesNoImagesProvider,
   AuthorArticlesWithImagesProvider,
-  ArticleProvider,
-  ArticleRelatedProvider
+  ArticleProvider
 };
